Cloud-Function/requirements.txt
google-auth==1.24.0
requests==2.25.1
requests_toolbelt==0.9.1
six==1.15.0